+++ ogcapi-drivers/src/postgres/feature.rs
@@ -1,5 +1,5 @@
 use ogcapi_types::{
-    common::{Authority, Bbox, Crs, Datetime, IntervalDatetime},
+    common::{Authority, Bbox, Crs, Datetime, Exception, IntervalDatetime},
     features::{Feature, FeatureCollection, Query},
 };
 
@@ -133,25 +133,39 @@ impl FeatureTransactions for Db {
 
     async fn list_items(
         &self,
-        collection: &str,
+        collection_id: &str,
         query: &Query,
     ) -> anyhow::Result<FeatureCollection> {
         let mut where_conditions = vec!["TRUE".to_owned()];
 
         // bbox
         if let Some(bbox) = query.bbox.as_ref() {
+            // crs
+            let bbox_crs = query.bbox_crs.clone().unwrap_or_else(|| match bbox {
+                Bbox::Bbox2D(_) => Crs::default2d(),
+                Bbox::Bbox3D(_) => Crs::default3d(),
+            });
+
             // coordinate system axis order (OGC and Postgis is lng, lat | EPSG is lat, lng)
-            let order = match query.bbox_crs.authority {
+            let order = match bbox_crs.authority {
                 Authority::OGC => [0, 1, 2, 3],
                 Authority::EPSG => [1, 0, 3, 2],
             };
 
-            let c = self.read_collection(collection).await?;
-            let storage_srid = c
-                .expect("collection exists")
-                .storage_crs
-                .unwrap_or_default()
-                .as_srid();
+            let collection = self.read_collection(collection_id).await?;
+            let storage_srid = match collection {
+                Some(collection) => match collection.storage_crs.map(|crs| crs.as_srid()) {
+                    Some(srid) => srid,
+                    None => {
+                        sqlx::query_scalar(&format!(
+                            "SELECT Find_SRID('items', '{collection_id}', 'geom')"
+                        ))
+                        .fetch_one(&self.pool)
+                        .await?
+                    }
+                },
+                None => return Err(Exception::new_from_status(404).into()),
+            };
 
             // TODO: handle antimeridian (lower > upper on axis 1)
             let intersection = match bbox {
@@ -161,7 +175,7 @@ impl FeatureTransactions for Db {
                     bbox[order[1]],
                     bbox[order[2]],
                     bbox[order[3]],
-                    query.bbox_crs.as_srid()
+                    bbox_crs.as_srid()
                 ),
                 Bbox::Bbox3D(bbox) => format!(
                     // FIXME: ensure proper height/box transformation handling
@@ -181,7 +195,7 @@ impl FeatureTransactions for Db {
                     x2 = bbox[order[2] + 1],
                     y2 = bbox[order[3] + 1],
                     z2 = bbox[5],
-                    srid = query.bbox_crs.as_srid()
+                    srid = bbox_crs.as_srid()
                 ),
             };
 
@@ -260,12 +274,19 @@ impl FeatureTransactions for Db {
         // count
         let number_matched: (i64,) = sqlx::query_as(&format!(
             r#"
-            SELECT count(*) FROM items."{collection}"
+            SELECT count(*) FROM items."{collection_id}"
             WHERE {conditions}
             "#,
         ))
         .fetch_one(&self.pool)
         .await?;
+
+        // srid
+        let srid = query
+            .crs
+            .as_ref()
+            .map(|crs| crs.as_srid())
+            .unwrap_or_else(|| Crs::default2d().as_srid());
 
         // fetch
         let features: Option<sqlx::types::Json<Vec<Feature>>> = sqlx::query_scalar(&format!(
@@ -273,7 +294,7 @@ impl FeatureTransactions for Db {
             SELECT array_to_json(array_agg(row_to_json(t)))
             FROM (
                 SELECT {ROWS}
-                FROM items."{collection}" items JOIN meta.collections meta
+                FROM items."{collection_id}" items JOIN meta.collections meta
                     ON items.collection = meta.id
                 WHERE {conditions}
                 LIMIT {}
@@ -285,7 +306,7 @@ impl FeatureTransactions for Db {
                 .map_or_else(|| String::from("NULL"), |l| l.to_string()),
             query.offset.unwrap_or(0)
         ))
-        .bind(query.crs.as_srid())
+        .bind(srid)
         .fetch_one(&self.pool)
         .await?;