+++ ogcapi-services/src/routes/tiles.rs
@@ -172,7 +172,7 @@ async fn tile_set() -> Result<Json<TileSet>> {
         data_type: DataType::Vector,
         tile_matrix_set_uri: Default::default(),
         tile_matrix_set_limits: Default::default(),
-        crs: TilesCrs::Simple(Crs::default().to_string()),
+        crs: TilesCrs::Simple(Crs::default2d().to_string()),
         epoch: Default::default(),
         links: Default::default(),
         layers: Default::default(),