+++ ogcapi-types/src/common/collection.rs
@@ -26,7 +26,8 @@ pub struct Collection {
     pub attribution: Option<String>,
     pub extent: Option<Extent>,
     /// An indicator about the type of the items in the collection.
-    pub item_type: Option<String>,
+    #[serde(default = "feature")]
+    pub item_type: String,
     /// The list of coordinate reference systems supported by the API; the first item is the default coordinate reference system.
     #[serde(default)]
     #[serde_as(as = "Vec<DisplayFromStr>")]
@@ -99,6 +100,10 @@ pub struct Collection {
     pub additional_properties: Map<String, Value>,
 }
 
+fn feature() -> String {
+    "feature".to_string()
+}
+
 #[cfg(feature = "stac")]
 fn collection() -> String {
     "Collection".to_string()
@@ -117,7 +122,7 @@ impl Default for Collection {
             attribution: Default::default(),
             extent: Default::default(),
             item_type: Default::default(),
-            crs: vec![Crs::default()],
+            crs: vec![Crs::default2d()],
             storage_crs: Default::default(),
             storage_crs_coordinate_epoch: Default::default(),
             links: Default::default(),