+++ ogcapi-types/src/common/query.rs
@@ -8,14 +8,14 @@ use crate::common::{Bbox, Crs, Datetime};
 #[derive(Deserialize, ToSchema, Debug, Clone)]
 #[serde(deny_unknown_fields, rename_all = "kebab-case")]
 pub struct Query {
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     #[serde_as(as = "Option<DisplayFromStr>")]
     pub bbox: Option<Bbox>,
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     #[serde_as(as = "Option<DisplayFromStr>")]
     #[schema(value_type = String)]
     pub bbox_crs: Option<Crs>,
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     #[serde_as(as = "Option<DisplayFromStr>")]
     #[schema(value_type = String)]
     pub datetime: Option<Datetime>,
@@ -28,7 +28,9 @@ pub struct Query {
 #[derive(Serialize, Deserialize, ToSchema, Debug, Clone, PartialEq, Eq)]
 pub struct LimitOffsetPagination {
     /// Amount of items to return
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub limit: Option<usize>,
     /// Offset into the items list
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub offset: Option<usize>,
 }